refactor: Heart와 Ellipsis View를 각각의 파일로 나눔
